# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Portal Berita",
        "description": "Portal Berita API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Local server"
        },
        {
            "url": "http://staging.example.com",
            "description": "Staging server"
        },
        {
            "url": "http://example.com",
            "description": "Production server"
        },
        {
            "url": ""
        }
    ],
    "paths": {
...
        "/supplier": {
            "get": {
                "tags": [
                    "Supplier"
                ],
                "summary": "List or search suppliers by name",
                "description": "Retrieve all suppliers or filter by partial match on 'nama' using query string",
                "operationId": "listOrSearchSupplier",
                "parameters": [
                    {
                        "name": "nama",
                        "in": "query",
                        "description": "Optional. Partial or full name of the supplier to search for",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "PT Maju"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of suppliers or search result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Supplier retrieved successfully."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "nama": {
                                                        "type": "string",
                                                        "example": "PT Maju"
                                                    },
                                                    "kontak": {
                                                        "type": "string",
                                                        "example": "0823456"
                                                    },
                                                    "alamat": {
                                                        "type": "string",
                                                        "example": "Jl. Merdeka"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No matching supplier found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 404
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "No matching supplier found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Supplier"
                ],
                "summary": "Create a new supplier",
                "description": "Add a new supplier",
                "operationId": "createSupplier",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama",
                                    "kontak",
                                    "alamat"
                                ],
                                "properties": {
                                    "nama": {
                                        "type": "string",
                                        "example": "PT Sejahtera"
                                    },
                                    "kontak": {
                                        "type": "string",
                                        "example": "08123456789"
                                    },
                                    "alamat": {
                                        "type": "string",
                                        "example": "Jl. Kebangsaan No.10"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Supplier created successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": 201,
                                    "message": "Supplier created successfully.",
                                    "data": {
                                        "id": 3,
                                        "nama": "PT Sejahtera",
                                        "kontak": "08123456789",
                                        "alamat": "Jl. Kebangsaan No.10"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/supplier/{id}": {
            "get": {
                "tags": [
                    "Supplier"
                ],
                "summary": "Get supplier by ID",
                "description": "Retrieve a single supplier by its ID",
                "operationId": "getSupplierById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Supplier found",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": 200,
                                    "message": "Supplier retrieved successfully.",
                                    "data": {
                                        "id": 1,
                                        "nama": "PT Maju",
                                        "kontak": "0823456",
                                        "alamat": "Jl. Merdeka"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Supplier not found",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": 404,
                                    "message": "Supplier not found.",
                                    "data": null
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Supplier"
                ],
                "summary": "Update a supplier",
                "description": "Update an existing supplier",
                "operationId": "updateSupplier",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama": {
                                        "type": "string",
                                        "example": "CV Makmur Jaya"
                                    },
                                    "kontak": {
                                        "type": "string",
                                        "example": "08987654321"
                                    },
                                    "alamat": {
                                        "type": "string",
                                        "example": "Jl. Industri No.5"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Supplier updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": 200,
                                    "message": "Supplier updated successfully.",
                                    "data": {
                                        "id": 1,
                                        "nama": "CV Makmur Jaya",
                                        "kontak": "08987654321",
                                        "alamat": "Jl. Industri No.5"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Supplier"
                ],
                "summary": "Delete a supplier",
                "description": "Delete a supplier by ID",
                "operationId": "deleteSupplier",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Supplier deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": 200,
                                    "message": "Supplier deleted successfully.",
                                    "data": null
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}